feat(config): make config hmr works both on Nuxt and Vite

+++ src/theme/output.ts
@@ -8,10 +8,9 @@ export function prepareOutputDir<UserOptions extends PinceauOptions = PinceauOpt
     outputDir = join(process.cwd(), 'node_modules/.vite/pinceau'),
   }: UserOptions,
 ) {
-  if (!existsSync(outputDir))
-    mkdirSync(outputDir, { recursive: true })
+  if (!existsSync(outputDir)) { mkdirSync(outputDir, { recursive: true }) }
 
-  stubOutputs(outputDir, true)
+  stubOutputs(outputDir, false)
 
   return outputDir
 }
@@ -28,10 +27,8 @@ export async function stubOutputs(buildPath: string, force = false) {
   for (const [file, stubbingFunction] of Object.entries(files)) {
     const path = join(buildPath, file)
 
-    if (force && existsSync(path))
-      rmSync(path)
+    if (force && existsSync(path)) { rmSync(path) }
 
-    if (!existsSync(path))
-      writeFileSync(path, stubbingFunction ? stubbingFunction({ tokens: {}, allTokens: [] } as any, {}) : '')
+    if (!existsSync(path)) { writeFileSync(path, stubbingFunction ? stubbingFunction({ tokens: {}, allTokens: [] } as any, {}) : '') }
   }
 }
